告警清零

+++ upload/frameworks/libs/src/obtain_file.cpp
@@ -59,10 +59,10 @@ uint32_t ObtainFile::GetDataAbilityFile(FILE **file, std::string &fileUri,
 {
     uint32_t ret = UPLOAD_ERRORCODE_NO_ERROR;
     FILE *filePtr = nullptr;
-    uint32_t fileLength = 0;
+    int32_t fileLength = 0;
 
     do {
-        uint32_t fd = fileAdapter_->DataAbilityOpenFile(fileUri, context);
+        int32_t fd = fileAdapter_->DataAbilityOpenFile(fileUri, context);
         if (fd == -1) {
             UPLOAD_HILOGE(UPLOAD_MODULE_FRAMEWORK, "ObtainFile::GetDataAbilityFile, open file error.");
             ret = UPLOAD_ERRORCODE_GET_FILE_ERROR;
@@ -87,7 +87,7 @@ uint32_t ObtainFile::GetDataAbilityFile(FILE **file, std::string &fileUri,
     } while (0);
 
     *file = filePtr;
-    fileSize = fileLength;
+    fileSize = static_cast<uint32_t>fileLength;
     return ret;
 }
